+++ database/migrations/2017_03_30_204311_create_itens_pedidos_table.php
@@ -21,7 +21,7 @@ class CreateItensPedidosTable extends Migration
             $table->integer('produto_id')->unsigned();
             $table->foreign('produto_id')->references('id')->on('produtos');
 
-            $table->integer('quantidade');
+            $table->float('quantidade');
 
             $table->float('preco');
